'Refactored by Sourcery'

+++ YukkiMusic/plugins/tools/queue.py
@@ -49,43 +49,37 @@ async def ping_com(client, message: Message, _):
                 await app.get_chat(chat_id)
             except:
                 return await message.reply_text(_["cplay_4"])
-    if await is_active_chat(chat_id):
-        got = db.get(chat_id)
-        if got:
-            send = await message.reply_text(_["queue_1"])
-            j = 0
-            msg = ""
-            for x in got:
-                j += 1
-                if j == 1:
-                    msg += f'**Currently Playing:**\n\n🏷 **Title:** `{x["title"]}`\n⏱ **Duration:** {x["dur"]}\n🎧 **Request By:** {x["by"]}\n\n'
-                elif j == 2:
-                    msg += f'**Queued:**\n🏷 **Title:** `{x["title"]}`\n⏱ **Duration:** {x["dur"]}\n🎧 **Request By:** {x["by"]}\n\n'
-                else:
-                    msg += f'🏷 **Title:** `{x["title"]}`\n⏱ **Duration:** {x["dur"]}\n🎧 **Request By:** {x["by"]}\n\n'
-            if "Queued" in msg:
-                link = await Yukkibin(msg)
-                lines = msg.count("\n")
-                if lines >= 55:
-                    car = os.linesep.join(msg.split(os.linesep)[:23])
-                else:
-                    return await send.edit_text(msg)
-                if "🏷" in car:
-                    car = car.replace("🏷", "")
-                if "⏱" in car:
-                    car = car.replace("⏱", "")
-                if "🎧" in car:
-                    car = car.replace("🎧", "")
-                carbon = await Carbon.generate(
-                    car, randint(100, 10000000)
-                )
-                await message.reply_photo(
-                    photo=carbon, caption=_["queue_3"].format(link)
-                )
-                await send.delete()
+    if await is_active_chat(chat_id) and (got := db.get(chat_id)):
+        send = await message.reply_text(_["queue_1"])
+        msg = ""
+        for j, x in enumerate(got, start=1):
+            if j == 1:
+                msg += f'**Currently Playing:**\n\n🏷 **Title:** `{x["title"]}`\n⏱ **Duration:** {x["dur"]}\n🎧 **Request By:** {x["by"]}\n\n'
+            elif j == 2:
+                msg += f'**Queued:**\n🏷 **Title:** `{x["title"]}`\n⏱ **Duration:** {x["dur"]}\n🎧 **Request By:** {x["by"]}\n\n'
             else:
-                await send.edit_text(msg)
+                msg += f'🏷 **Title:** `{x["title"]}`\n⏱ **Duration:** {x["dur"]}\n🎧 **Request By:** {x["by"]}\n\n'
+        if "Queued" in msg:
+            link = await Yukkibin(msg)
+            lines = msg.count("\n")
+            if lines >= 55:
+                car = os.linesep.join(msg.split(os.linesep)[:23])
+            else:
+                return await send.edit_text(msg)
+            if "🏷" in car:
+                car = car.replace("🏷", "")
+            if "⏱" in car:
+                car = car.replace("⏱", "")
+            if "🎧" in car:
+                car = car.replace("🎧", "")
+            carbon = await Carbon.generate(
+                car, randint(100, 10000000)
+            )
+            await message.reply_photo(
+                photo=carbon, caption=_["queue_3"].format(link)
+            )
+            await send.delete()
         else:
-            await message.reply_text(_["queue_2"])
+            await send.edit_text(msg)
     else:
         await message.reply_text(_["queue_2"])
